requirements.txt
Flask
gunicorn
sqlalchemy
firebase_admin
pymysql
python-dotenv
requests
Authlib